requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
backports.functools-lru-cache==1.6.5
comm==0.1.4
debugpy==1.6.8
decorator==5.1.1
executing==1.2.0
feedparser==6.0.8
greenlet==2.0.2
importlib-metadata==6.8.0
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==23.2.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
python-decouple==3.8
pyzmq==25.1.1
setuptools==68.0.0
sgmllib3k==1.0.0
six==1.16.0
SQLAlchemy==2.0.20
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
wcwidth==0.2.6
wheel==0.38.4
zipp==3.16.2
